pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-flow-next"
version = "0.1.0"
description = "Three-layer AgentFlow runtime and governance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1",
  "pydantic>=2.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

[project.scripts]
agent-flow = "agent_flow.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
